Still connecting front end and back end
The ID is now passed as a prop from the database components; plan to condense the database components down into 1, and pass this way to get closer to DRY

The database component also handles sending to the backend now via a function prop.

+++ src/main/java/com/twag/book_management_app/repository/BookRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.twag.book_management_app.model.Book; // FIXME: this will likely need to be a 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 }
